+++ corebehrt/main/create_outcomes.py
@@ -10,8 +10,9 @@ from tqdm import tqdm
 from corebehrt.modules.setup.config import load_config
 from corebehrt.modules.monitoring.logger import TqdmToLogger
 from corebehrt.modules.setup.directory import DirectoryPreparer, get_args
-from corebehrt.data.concept_loader import ConceptLoaderLarge
+from corebehrt.modules.features.loader import ConceptLoaderLarge
 from corebehrt.modules.cohort_handling.outcomes import OutcomeMaker
+
 
 CONFIG_PATH = "./corebehrt/configs/outcomes_test.yaml"
